apps/workers/publisher/pyproject.toml
[project]
name = "secai-radar-publisher"
version = "1.0.0"
description = "Publisher worker for atomic dataset swap"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
